backend/requirements.txt
# backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis and Celery - Fixed version conflict
redis==4.6.0
celery==5.3.4
flower==2.0.1

# Storage
minio==7.2.0

# API enhancements
fastapi-limiter==0.1.5
slowapi==0.1.9

# Internationalization
babel==2.13.1

# Email
emails==0.6
jinja2==3.1.2

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Email validation for Pydantic
email-validator>=2.0.0